feat: Fix compile error mismatched types
The problem is, as the compiler pointed out, there were two versions of
egui v0.23 and v0.22. I fixed this by forking.
  https://github.com/Gip-Gip/egui-plotter to
  https://github.com/winksaville/egui-plotter

And then changing eframe and egui version in `Cargo.toml` to 0.23 and
creating a branch ,`update-eframe-egui-to-0.23`, on my fork:
  https://github.com/winksaville/egui-plotter/tree/update-eframe-egui-to-0.23

I then changed the `Cargo.toml` here to point to that fork and now it's
working.